# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WildGuard proposed backend API",
    "version": "1.0.0",
    "description": "Original proposed full-system contract. Only milestone 1 is implemented; use implemented-openapi.json for its current envelope and APIs. Other modules remain future work."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Planned full-system API; see implemented-openapi.json for the implemented subset."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/patrol-routes/{id}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Get a patrol route",
        "operationId": "get_Reference_6",
        "x-roles": [
          "PARK_MANAGER",
          "RANGER"
        ],
        "description": "Park-scoped route geometry and metadata.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatrolRoute"
                },
                "example": {
                  "id": "route-yala-b1",
                  "parkId": "park-yala",
                  "areaId": "area-b1",
                  "name": "Block 1 wildlife trail",
                  "plannedDistanceMeters": 8500,
                  "pathPoints": [
                    {
                      "latitude": 6.35,
                      "longitude": 81.5
                    },
                    {
                      "latitude": 6.351,
                      "longitude": 81.501
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": ".*\\S.*"
            },
            "example": "route-yala-b1"
          }
        ]
      }
    },
...
    "/patrols/{id}": {
      "put": {
        "tags": [
          "Patrols"
        ],
        "summary": "Submit a completed patrol",
        "operationId": "put_Patrols_9",
        "x-roles": [
          "RANGER"
        ],
        "description": "Immutable completed snapshot; active recording remains local. Validate assignment ownership, one patrol per assignment, ordered timestamps and location/array limits. Compute distance from continuous GPS segments, ignoring gaps over 120 seconds and fixes with reported accuracy worse than 100 meters; manual waypoints do not add travel distance. Identical replay is safe; changed replay or reused assignment is 409.",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatrolDetail"
                },
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "assignmentId": "11111111-1111-4111-8111-111111111111",
                  "routeId": "route-yala-b1",
                  "parkId": "park-yala",
                  "rangerId": "usr-ranger-1",
                  "status": "COMPLETED",
                  "startedAt": "2026-10-08T02:45:00Z",
                  "endedAt": "2026-10-08T06:40:00Z",
                  "recordedDistanceMeters": 15.7,
                  "waypointCount": 1,
                  "observationCount": 1,
                  "createdAt": "2026-10-08T06:41:00Z",
                  "trackPoints": [
                    {
                      "latitude": 6.35,
                      "longitude": 81.5,
                      "recordedAt": "2026-10-08T02:45:00Z",
                      "accuracyMeters": 8
                    },
                    {
                      "latitude": 6.3501,
                      "longitude": 81.5001,
                      "recordedAt": "2026-10-08T02:45:30Z",
                      "accuracyMeters": 8
                    }
                  ],
                  "waypoints": [
                    {
                      "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                      "label": "Fresh footprints",
                      "recordedAt": "2026-10-08T04:00:00Z",
                      "location": {
                        "latitude": 6.35,
                        "longitude": 81.5,
                        "source": "GPS",
                        "accuracyMeters": 8
                      }
                    }
                  ],
                  "observations": [
                    {
                      "id": "bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb",
                      "text": "Fresh elephant footprints near the trail.",
                      "observedAt": "2026-10-08T04:00:00Z",
                      "location": {
                        "latitude": 6.35,
                        "longitude": 81.5,
                        "source": "GPS",
                        "accuracyMeters": 8
                      }
                    }
                  ]
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URI of the created resource.",
                "schema": {
                  "type": "string"
                },
                "example": "/api/v1/patrols/22222222-2222-4222-8222-222222222222"
              }
            }
          },
          "200": {
            "description": "Identical replay returns the existing record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatrolDetail"
                },
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "assignmentId": "11111111-1111-4111-8111-111111111111",
                  "routeId": "route-yala-b1",
                  "parkId": "park-yala",
                  "rangerId": "usr-ranger-1",
                  "status": "COMPLETED",
                  "startedAt": "2026-10-08T02:45:00Z",
                  "endedAt": "2026-10-08T06:40:00Z",
                  "recordedDistanceMeters": 15.7,
                  "waypointCount": 1,
                  "observationCount": 1,
                  "createdAt": "2026-10-08T06:41:00Z",
                  "trackPoints": [
                    {
                      "latitude": 6.35,
                      "longitude": 81.5,
                      "recordedAt": "2026-10-08T02:45:00Z",
                      "accuracyMeters": 8
                    },
                    {
                      "latitude": 6.3501,
                      "longitude": 81.5001,
                      "recordedAt": "2026-10-08T02:45:30Z",
                      "accuracyMeters": 8
                    }
                  ],
                  "waypoints": [
                    {
                      "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                      "label": "Fresh footprints",
                      "recordedAt": "2026-10-08T04:00:00Z",
                      "location": {
                        "latitude": 6.35,
                        "longitude": 81.5,
                        "source": "GPS",
                        "accuracyMeters": 8
                      }
                    }
                  ],
                  "observations": [
                    {
                      "id": "bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb",
                      "text": "Fresh elephant footprints near the trail.",
                      "observedAt": "2026-10-08T04:00:00Z",
                      "location": {
                        "latitude": 6.35,
                        "longitude": 81.5,
                        "source": "GPS",
                        "accuracyMeters": 8
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "22222222-2222-4222-8222-222222222222"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatrolRequest"
              },
              "example": {
                "assignmentId": "11111111-1111-4111-8111-111111111111",
                "startedAt": "2026-10-08T02:45:00Z",
                "endedAt": "2026-10-08T06:40:00Z",
                "trackPoints": [
                  {
                    "latitude": 6.35,
                    "longitude": 81.5,
                    "recordedAt": "2026-10-08T02:45:00Z",
                    "accuracyMeters": 8
                  },
                  {
                    "latitude": 6.3501,
                    "longitude": 81.5001,
                    "recordedAt": "2026-10-08T02:45:30Z",
                    "accuracyMeters": 8
                  }
                ],
                "waypoints": [
                  {
                    "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "label": "Fresh footprints",
                    "recordedAt": "2026-10-08T04:00:00Z",
                    "location": {
                      "latitude": 6.35,
                      "longitude": 81.5,
                      "source": "GPS",
                      "accuracyMeters": 8
                    }
                  }
                ],
                "observations": [
                  {
                    "id": "bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb",
                    "text": "Fresh elephant footprints near the trail.",
                    "observedAt": "2026-10-08T04:00:00Z",
                    "location": {
                      "latitude": 6.35,
                      "longitude": 81.5,
                      "source": "GPS",
                      "accuracyMeters": 8
                    }
                  }
                ]
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Patrols"
        ],
        "summary": "Get completed patrol and recorded path",
        "operationId": "get_Patrols_11",
        "x-roles": [
          "PARK_MANAGER",
          "RANGER"
        ],
        "description": "Owner ranger or authorized park manager. Includes stored waypoints, observations and track points.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatrolDetail"
                },
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "assignmentId": "11111111-1111-4111-8111-111111111111",
                  "routeId": "route-yala-b1",
                  "parkId": "park-yala",
                  "rangerId": "usr-ranger-1",
                  "status": "COMPLETED",
                  "startedAt": "2026-10-08T02:45:00Z",
                  "endedAt": "2026-10-08T06:40:00Z",
                  "recordedDistanceMeters": 15.7,
                  "waypointCount": 1,
                  "observationCount": 1,
                  "createdAt": "2026-10-08T06:41:00Z",
                  "trackPoints": [
                    {
                      "latitude": 6.35,
                      "longitude": 81.5,
                      "recordedAt": "2026-10-08T02:45:00Z",
                      "accuracyMeters": 8
                    },
                    {
                      "latitude": 6.3501,
                      "longitude": 81.5001,
                      "recordedAt": "2026-10-08T02:45:30Z",
                      "accuracyMeters": 8
                    }
                  ],
                  "waypoints": [
                    {
                      "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                      "label": "Fresh footprints",
                      "recordedAt": "2026-10-08T04:00:00Z",
                      "location": {
                        "latitude": 6.35,
                        "longitude": 81.5,
                        "source": "GPS",
                        "accuracyMeters": 8
                      }
                    }
                  ],
                  "observations": [
                    {
                      "id": "bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb",
                      "text": "Fresh elephant footprints near the trail.",
                      "observedAt": "2026-10-08T04:00:00Z",
                      "location": {
                        "latitude": 6.35,
                        "longitude": 81.5,
                        "source": "GPS",
                        "accuracyMeters": 8
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "22222222-2222-4222-8222-222222222222"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "code",
          "message",
          "path"
        ],
        "additionalProperties": false
      },
      "Location": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "source": {
            "type": "string",
            "enum": [
              "GPS",
              "MANUAL"
            ]
          },
          "accuracyMeters": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          }
        },
        "required": [
          "latitude",
          "longitude",
          "source"
        ],
        "additionalProperties": false
      },
...
      "RoutePoint": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "additionalProperties": false
      },
      "PatrolRoute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "parkId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "areaId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150,
            "pattern": ".*\\S.*"
          },
          "plannedDistanceMeters": {
            "type": "number",
            "minimum": 0
          },
          "pathPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutePoint"
            },
            "minItems": 1,
            "maxItems": 1000
          }
        },
        "required": [
          "id",
          "parkId",
          "areaId",
          "name",
          "plannedDistanceMeters",
          "pathPoints"
        ],
        "additionalProperties": false
      },
...
      "TrackPoint": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          },
          "accuracyMeters": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          }
        },
        "required": [
          "latitude",
          "longitude",
          "recordedAt"
        ],
        "additionalProperties": false
      },
      "Waypoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": ".*\\S.*"
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "notes": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "pattern": ".*\\S.*"
          }
        },
        "required": [
          "id",
          "label",
          "recordedAt",
          "location"
        ],
        "additionalProperties": false
      },
      "Observation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "pattern": ".*\\S.*"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          }
        },
        "required": [
          "id",
          "text",
          "observedAt",
          "location"
        ],
        "additionalProperties": false
      },
      "PatrolRequest": {
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string",
            "format": "uuid"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          },
          "trackPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackPoint"
            },
            "maxItems": 10000
          },
          "waypoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Waypoint"
            },
            "maxItems": 100
          },
          "observations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "maxItems": 100
          }
        },
        "required": [
          "assignmentId",
          "startedAt",
          "endedAt",
          "trackPoints",
          "waypoints",
          "observations"
        ],
        "additionalProperties": false
      },
...
      "PatrolDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "assignmentId": {
            "type": "string",
            "format": "uuid"
          },
          "routeId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "parkId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "rangerId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          },
          "recordedDistanceMeters": {
            "type": "number",
            "minimum": 0
          },
          "waypointCount": {
            "type": "integer",
            "minimum": 0
          },
          "observationCount": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "trackPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackPoint"
            },
            "maxItems": 10000
          },
          "waypoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Waypoint"
            },
            "maxItems": 100
          },
          "observations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "maxItems": 100
          }
        },
        "required": [
          "id",
          "assignmentId",
          "routeId",
          "parkId",
          "rangerId",
          "status",
          "startedAt",
          "endedAt",
          "recordedDistanceMeters",
          "waypointCount",
          "observationCount",
          "createdAt",
          "trackPoints",
          "waypoints",
          "observations"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Invalid request or validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 400,
              "code": "VALIDATION_FAILED",
              "message": "Invalid request or validation failed.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error401": {
        "description": "Missing/invalid token, inactive account or invalid login.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 401,
              "code": "UNAUTHORIZED",
              "message": "Missing/invalid token, inactive account or invalid login.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error403": {
        "description": "Role or requested park is forbidden.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 403,
              "code": "FORBIDDEN",
              "message": "Role or requested park is forbidden.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error404": {
        "description": "Resource not found or not accessible.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 404,
              "code": "NOT_FOUND",
              "message": "Resource not found or not accessible.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error409": {
        "description": "Conflicting state or resource ID reused with different data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 409,
              "code": "STATE_CONFLICT",
              "message": "Conflicting state or resource ID reused with different data.",
              "path": "/api/v1/example"
            }
          }
        }
      },
...
      "Error500": {
        "description": "Unexpected server error; no false success acknowledgement.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 500,
              "code": "INTERNAL_ERROR",
              "message": "Unexpected server error; no false success acknowledgement.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error503": {
        "description": "Database/storage temporarily unavailable; retry later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 503,
              "code": "SERVICE_UNAVAILABLE",
              "message": "Database/storage temporarily unavailable; retry later.",
              "path": "/api/v1/example"
            }
          }
        }
      }
    }
  }
}